moved setup to subpackage
