+++ cms_perf/setup/cli.py
@@ -1,10 +1,10 @@
 import argparse
 
-from . import cli_parser
-from . import __version__ as lib_version
+from ..setup import cli_parser
+from .. import __version__ as lib_version
 
 # ensure sensors are loaded
-from .sensors import sensor, xrd_load, net_load  # noqa
+from ..sensors import sensor, xrd_load, net_load  # noqa
 
 
 INTERVAL_UNITS = {"": 1, "s": 1, "m": 60, "h": 60 * 60}
